Relocate HGCal Hit Calibration code
