--- Validation/HGCalHitValidation/python/hgcalHitValidation_cfi.py
@@ -1,5 +1,0 @@
-import FWCore.ParameterSet.Config as cms
-
-from Validation.HGCalHitValidation.hgcalHitCalibration_cfi import hgcalHitCalibration
-
-hgcalHitValidationSequece = cms.Sequence(hgcalHitCalibration)
