+++ Validation/HGCalValidation/test/standalone_fromRECO_HARVESTING.py
